chore(chat): implement chat page UI with slash command panel and workspace surface
