+++ ui/lib/features/home/pages/chat/chat_page_conversation_flow.dart
@@ -1,0 +1,705 @@
+part of 'chat_page.dart';
+
+mixin _ChatPageConversationFlowMixin on _ChatPageStateBase {
+  @override
+  void _syncRuntimeSnapshotForMode(
+    ChatPageMode mode, {
+    ConversationModel? conversation,
+    List<ChatMessageModel>? messages,
+  }) {
+    final conversationId = _currentConversationIdByMode[mode];
+    if (conversationId == null) return;
+    final runtime = _runtimeCoordinator.runtimeFor(
+      conversationId: conversationId,
+      mode: _modeKey(mode),
+    );
+    _runtimeCoordinator.replaceConversationSnapshot(
+      conversationId: conversationId,
+      mode: _modeKey(mode),
+      messages: List<ChatMessageModel>.from(
+        messages ?? runtime?.messages ?? _messagesByMode[mode]!,
+      ),
+      conversation:
+          conversation ??
+          runtime?.conversation ??
+          _currentConversationByMode[mode],
+      isAiResponding:
+          runtime?.isAiResponding ?? (_isAiRespondingByMode[mode] ?? false),
+      isCheckingExecutableTask:
+          runtime?.isCheckingExecutableTask ??
+          (_isCheckingExecutableTaskByMode[mode] ?? false),
+      isSubmittingVlmReply:
+          runtime?.isSubmittingVlmReply ??
+          (_isSubmittingVlmReplyByMode[mode] ?? false),
+      vlmInfoQuestion: runtime?.vlmInfoQuestion ?? _vlmInfoQuestionByMode[mode],
+      currentAiMessages: Map<String, String>.from(
+        runtime?.currentAiMessages ?? _currentAiMessagesByMode[mode]!,
+      ),
+      deepThinkingContent:
+          runtime?.deepThinkingContent ??
+          (_deepThinkingContentByMode[mode] ?? ''),
+      isDeepThinking:
+          runtime?.isDeepThinking ?? (_isDeepThinkingByMode[mode] ?? false),
+      currentDispatchTaskId:
+          runtime?.currentDispatchTaskId ?? _currentDispatchTaskIdByMode[mode],
+      currentThinkingStage:
+          runtime?.currentThinkingStage ??
+          (_currentThinkingStageByMode[mode] ?? 1),
+      isInputAreaVisible:
+          runtime?.isInputAreaVisible ??
+          (_isInputAreaVisibleByMode[mode] ?? true),
+      isExecutingTask:
+          runtime?.isExecutingTask ?? (_isExecutingTaskByMode[mode] ?? false),
+      lastAgentTaskId: runtime?.lastAgentTaskId,
+      activeToolCardId: runtime?.activeToolCardId,
+      activeThinkingCardId: runtime?.activeThinkingCardId,
+      pendingAgentTextTaskId: runtime?.pendingAgentTextTaskId,
+      pendingThinkingRoundSplit: runtime?.pendingThinkingRoundSplit ?? false,
+      toolCardSequence: runtime?.toolCardSequence ?? 0,
+      thinkingRound: runtime?.thinkingRound ?? 0,
+    );
+  }
+
+  @override
+  Future<void> _ensureActiveConversationReadyForStreaming() async {
+    if (_currentConversationId == null) {
+      await persistConversationSnapshot(
+        generateSummary: false,
+        markComplete: false,
+      );
+    }
+    if (_currentConversationId == null) {
+      throw StateError('conversationId is not ready');
+    }
+    _syncRuntimeSnapshotForMode(_activeMode);
+  }
+
+  @override
+  void _registerActiveTaskBinding(String taskId) {
+    final conversationId = _currentConversationId;
+    if (conversationId == null) return;
+    _runtimeCoordinator.registerTask(
+      taskId: taskId,
+      conversationId: conversationId,
+      mode: _modeKey(_activeMode),
+    );
+  }
+
+  @override
+  void _createThinkingCard(
+    String taskID, {
+    String? cardId,
+    String? thinkingContent,
+    bool? isLoading,
+    int? stage,
+  }) {
+    final loadingIndex = _messages.indexWhere((msg) => msg.id == taskID);
+    if (loadingIndex != -1) {
+      setState(() => _messages.removeAt(loadingIndex));
+    }
+
+    final startTime = DateTime.now().millisecondsSinceEpoch;
+    final thinkingCardId = cardId ?? '$taskID-thinking';
+    final cardData = {
+      'type': 'deep_thinking',
+      'isLoading': isLoading ?? _isDeepThinking,
+      'thinkingContent': thinkingContent ?? '',
+      'stage': stage ?? _currentThinkingStage,
+      'taskID': taskID,
+      'startTime': startTime,
+      'endTime': null,
+    };
+
+    setState(() {
+      _messages.removeWhere((msg) => msg.id == thinkingCardId);
+      _messages.insert(
+        0,
+        ChatMessageModel(
+          id: thinkingCardId,
+          type: 2,
+          user: 3,
+          content: {'cardData': cardData, 'id': thinkingCardId},
+        ),
+      );
+    });
+  }
+
+  @override
+  void _updateThinkingCard(
+    String taskID, {
+    String? cardId,
+    String? thinkingContent,
+    bool? isLoading,
+    int? stage,
+    bool lockCompleted = true,
+  }) {
+    final thinkingCardId = cardId ?? '$taskID-thinking';
+    final index = _messages.indexWhere((msg) => msg.id == thinkingCardId);
+    if (index == -1) return;
+
+    setState(() {
+      final existing = _messages[index];
+      final content = Map<String, dynamic>.from(existing.content ?? {});
+      final cardData = Map<String, dynamic>.from(content['cardData'] ?? {});
+
+      final currentStage = cardData['stage'] as int? ?? 1;
+      final targetStage = stage ?? _currentThinkingStage;
+      final newStage = (lockCompleted && currentStage == 4) ? 4 : targetStage;
+
+      final startTime = cardData['startTime'] as int?;
+      int? endTime = cardData['endTime'] as int?;
+      if (newStage == 4 && endTime == null) {
+        endTime = DateTime.now().millisecondsSinceEpoch;
+      }
+
+      cardData['thinkingContent'] = thinkingContent ?? _deepThinkingContent;
+      cardData['isLoading'] = isLoading ?? _isDeepThinking;
+      cardData['stage'] = newStage;
+      cardData['taskID'] = taskID;
+      cardData['startTime'] = startTime;
+      cardData['endTime'] = endTime;
+
+      content['cardData'] = cardData;
+      _messages[index] = existing.copyWith(content: content);
+    });
+  }
+
+  @override
+  Future<void> _pickAttachments() async {
+    try {
+      final result = await FilePicker.platform.pickFiles(
+        allowMultiple: true,
+        type: FileType.any,
+      );
+      if (result == null || result.files.isEmpty || !mounted) return;
+
+      setState(() {
+        for (final file in result.files) {
+          final path = file.path;
+          if (path == null || path.isEmpty) continue;
+          final exists = _pendingAttachments.any((item) => item.path == path);
+          if (exists) continue;
+          final displayName = (file.name.trim().isNotEmpty)
+              ? file.name.trim()
+              : _fileNameFromPath(path);
+          final extension = (file.extension ?? '').toLowerCase();
+          final mimeType = _mimeTypeFromExtension(path, extension: extension);
+          final isImage = _isImageFilePath(path, mimeType: mimeType);
+          _pendingAttachments.add(
+            ChatInputAttachment(
+              id: '${path}_${DateTime.now().microsecondsSinceEpoch}',
+              name: displayName,
+              path: path,
+              size: file.size > 0 ? file.size : null,
+              mimeType: mimeType,
+              isImage: isImage,
+            ),
+          );
+        }
+      });
+    } catch (e) {
+      _showSnackBar('添加附件失败：$e');
+    }
+  }
+
+  @override
+  void _removePendingAttachment(String id) {
+    if (!mounted) return;
+    setState(() {
+      _pendingAttachments.removeWhere((item) => item.id == id);
+    });
+  }
+
+  @override
+  String _fileNameFromPath(String path) {
+    final normalized = path.replaceAll('\\', '/');
+    final segments = normalized.split('/');
+    if (segments.isEmpty) return path;
+    return segments.last.isEmpty ? path : segments.last;
+  }
+
+  @override
+  bool _isImageFilePath(String path, {String? mimeType}) {
+    final normalizedMime = mimeType?.trim().toLowerCase();
+    if (normalizedMime != null && normalizedMime.startsWith('image/')) {
+      return true;
+    }
+    final lowerPath = path.toLowerCase();
+    return lowerPath.endsWith('.png') ||
+        lowerPath.endsWith('.jpg') ||
+        lowerPath.endsWith('.jpeg') ||
+        lowerPath.endsWith('.webp') ||
+        lowerPath.endsWith('.gif') ||
+        lowerPath.endsWith('.bmp') ||
+        lowerPath.endsWith('.heic') ||
+        lowerPath.endsWith('.heif');
+  }
+
+  @override
+  String? _mimeTypeFromExtension(String path, {String extension = ''}) {
+    final ext = extension.isNotEmpty
+        ? extension
+        : _fileNameFromPath(path).split('.').last.toLowerCase();
+    switch (ext) {
+      case 'png':
+        return 'image/png';
+      case 'jpg':
+      case 'jpeg':
+        return 'image/jpeg';
+      case 'gif':
+        return 'image/gif';
+      case 'webp':
+        return 'image/webp';
+      case 'bmp':
+        return 'image/bmp';
+      case 'heic':
+        return 'image/heic';
+      case 'heif':
+        return 'image/heif';
+      case 'pdf':
+        return 'application/pdf';
+      case 'txt':
+        return 'text/plain';
+      case 'md':
+        return 'text/markdown';
+      default:
+        return null;
+    }
+  }
+
+  @override
+  void _showSnackBar(String message) {
+    if (!mounted) return;
+    final messenger = ScaffoldMessenger.maybeOf(context);
+    messenger?.hideCurrentSnackBar();
+    messenger?.showSnackBar(
+      SnackBar(
+        content: Text(message),
+        duration: const Duration(milliseconds: 1200),
+        behavior: SnackBarBehavior.floating,
+      ),
+    );
+  }
+
+  @override
+  Future<void> _sendMessage({String? text}) async {
+    final messageText = (text ?? _messageController.text).trim();
+    final hasAttachments = _pendingAttachments.isNotEmpty;
+    if ((messageText.isEmpty && !hasAttachments) || _isAiResponding) return;
+
+    if (messageText.isNotEmpty) {
+      final handledSlash = await _tryHandleSlashCommand(messageText);
+      if (handledSlash) return;
+    }
+
+    if (_isOpenClawSurface && _openClawBaseUrl.trim().isEmpty) {
+      _showSnackBar('请先使用 /deploy 或 /openclaw 完成配置');
+      unawaited(_showOpenClawDeployPanel());
+      return;
+    }
+
+    final attachments = _pendingAttachments
+        .map((item) => item.toMap())
+        .toList();
+    if (attachments.isNotEmpty && mounted) {
+      setState(() => _pendingAttachments.clear());
+    }
+
+    _inputFocusNode.unfocus();
+    final messageIds = addUserMessage(messageText, attachments: attachments);
+
+    if (_isOpenClawSurface) {
+      await _sendChatMessage(messageIds.aiMessageId);
+      return;
+    }
+
+    try {
+      await _ensureActiveConversationReadyForStreaming();
+    } catch (_) {
+      if (mounted) {
+        handleAgentError('Conversation setup failed. Please retry.');
+      }
+      return;
+    }
+
+    final handled = await _handleExecutableTaskFlow(
+      messageIds.aiMessageId,
+      messageIds.userMessageId,
+    );
+    if (!handled &&
+        mounted &&
+        _currentDispatchTaskId == messageIds.aiMessageId) {
+      handleAgentError('统一 Agent 启动失败，请检查模型提供商与场景模型配置。');
+    }
+  }
+
+  @override
+  Future<void> _sendChatMessage(String aiMessageId) async {
+    if (!_isOpenClawSurface) {
+      handleAgentError('统一 Agent 已启用，旧聊天链路已移除，请检查配置后重试。');
+      return;
+    }
+    final history = buildConversationHistory();
+    final openClawConfig = {
+      'baseUrl': _openClawBaseUrl,
+      if (_openClawToken.isNotEmpty) 'token': _openClawToken,
+      if (_openClawUserId.isNotEmpty) 'userId': _openClawUserId,
+      if (_openClawUserId.isNotEmpty)
+        'sessionKey': 'openclaw:${_openClawUserId.trim()}',
+    };
+    try {
+      await _ensureActiveConversationReadyForStreaming();
+    } catch (_) {
+      if (mounted) {
+        handleAgentError('Conversation setup failed. Please retry.');
+      }
+      return;
+    }
+    _showOpenClawWaitingCard(aiMessageId);
+    _syncRuntimeSnapshotForMode(_activeMode);
+    _registerActiveTaskBinding(aiMessageId);
+    final success = await AssistsMessageService.createChatTask(
+      aiMessageId,
+      history,
+      provider: 'openclaw',
+      openClawConfig: openClawConfig,
+    );
+    if (success) return;
+    _runtimeCoordinator.unregisterTask(aiMessageId);
+
+    try {
+      throw Exception('createChatTask returned false');
+    } catch (error) {
+      if (!mounted) return;
+      final errorId = DateTime.now().millisecondsSinceEpoch.toString();
+      _removeOpenClawWaitingCard(aiMessageId);
+      setState(() {
+        _isAiResponding = false;
+        removeLatestLoadingIfExists();
+        _messages.insert(
+          0,
+          ChatMessageModel(
+            id: errorId,
+            type: 1,
+            user: 2,
+            content: {'text': '抱歉，发送消息失败：$error', 'id': errorId},
+          ),
+        );
+      });
+    }
+  }
+
+  @override
+  Future<bool> _handleExecutableTaskFlow(
+    String aiMessageId,
+    String userMessageId,
+  ) async {
+    _isCheckingExecutableTask = true;
+    try {
+      return await _tryAgentFlow(aiMessageId, userMessageId);
+    } finally {
+      _isCheckingExecutableTask = false;
+    }
+  }
+
+  @override
+  Future<bool> _tryAgentFlow(String aiMessageId, String userMessageId) async {
+    try {
+      _currentDispatchTaskId = aiMessageId;
+      _deepThinkingContent = '';
+      _isDeepThinking = false;
+      _currentThinkingStage = 1;
+
+      createThinkingCard(aiMessageId);
+      _syncRuntimeSnapshotForMode(_activeMode);
+      _registerActiveTaskBinding(aiMessageId);
+
+      final userMessage = latestUserUtterance();
+      final history = _historyBeforeLatestUser(buildConversationHistory());
+      final attachments = await _latestUserAttachments();
+
+      final success = await AssistsMessageService.createAgentTask(
+        taskId: aiMessageId,
+        userMessage: userMessage,
+        conversationHistory: history,
+        attachments: attachments,
+        conversationId: _currentConversationId,
+        modelOverride: _buildAgentModelOverridePayload(),
+      );
+      if (!success) {
+        _runtimeCoordinator.unregisterTask(aiMessageId);
+      }
+
+      return success;
+    } catch (e) {
+      _runtimeCoordinator.unregisterTask(aiMessageId);
+      debugPrint('Agent flow error: $e');
+      return false;
+    }
+  }
+
+  @override
+  List<Map<String, dynamic>> _historyBeforeLatestUser(
+    List<Map<String, dynamic>> history,
+  ) {
+    if (history.isEmpty) return history;
+    final normalized = List<Map<String, dynamic>>.from(history);
+    final last = normalized.last;
+    if ((last['role'] as String?) == 'user') {
+      normalized.removeLast();
+    }
+    return normalized;
+  }
+
+  @override
+  Future<List<Map<String, dynamic>>> _latestUserAttachments() async {
+    for (final message in _messages) {
+      if (message.user != 1) continue;
+      final raw = message.content?['attachments'];
+      if (raw is! List) return const [];
+      final normalized = raw
+          .whereType<Map>()
+          .map((item) => item.map((k, v) => MapEntry(k.toString(), v)))
+          .toList();
+      for (final item in normalized) {
+        if (!_isImageAttachmentMap(item)) continue;
+        final dataUrl = await _resolveImageDataUrl(item);
+        if (dataUrl.isNotEmpty) {
+          item['dataUrl'] = dataUrl;
+        }
+      }
+      return normalized;
+    }
+    return const [];
+  }
+
+  @override
+  bool _isImageAttachmentMap(Map<String, dynamic> item) {
+    final explicitFlag = item['isImage'];
+    if (explicitFlag is bool && explicitFlag) return true;
+    final mimeType = (item['mimeType'] as String? ?? '').toLowerCase();
+    if (mimeType.startsWith('image/')) return true;
+    final path = (item['path'] as String? ?? '').toLowerCase();
+    final url = (item['url'] as String? ?? '').toLowerCase();
+    return path.endsWith('.png') ||
+        path.endsWith('.jpg') ||
+        path.endsWith('.jpeg') ||
+        path.endsWith('.webp') ||
+        path.endsWith('.gif') ||
+        path.endsWith('.bmp') ||
+        path.endsWith('.heic') ||
+        path.endsWith('.heif') ||
+        url.endsWith('.png') ||
+        url.endsWith('.jpg') ||
+        url.endsWith('.jpeg') ||
+        url.endsWith('.webp') ||
+        url.endsWith('.gif');
+  }
+
+  @override
+  Future<String> _resolveImageDataUrl(Map<String, dynamic> item) async {
+    final existingDataUrl = (item['dataUrl'] as String? ?? '').trim();
+    if (existingDataUrl.startsWith('data:')) {
+      return existingDataUrl;
+    }
+
+    final existingUrl = (item['url'] as String? ?? '').trim();
+    if (existingUrl.startsWith('data:')) {
+      return existingUrl;
+    }
+    if (existingUrl.startsWith('http://') ||
+        existingUrl.startsWith('https://')) {
+      return existingUrl;
+    }
+
+    final path = (item['path'] as String? ?? '').trim();
+    if (path.isEmpty) return '';
+    final file = File(path);
+    if (!await file.exists()) return '';
+    try {
+      final bytes = await file.readAsBytes();
+      if (bytes.isEmpty) return '';
+      final mimeType = ((item['mimeType'] as String?) ?? '')
+          .trim()
+          .toLowerCase();
+      final resolvedMime = mimeType.startsWith('image/')
+          ? mimeType
+          : _mimeTypeFromExtension(path) ?? 'image/png';
+      return 'data:$resolvedMime;base64,${base64Encode(bytes)}';
+    } catch (_) {
+      return '';
+    }
+  }
+
+  @override
+  void _onCancelTask() {
+    try {
+      if (_currentDispatchTaskId != null ||
+          _isCheckingExecutableTask ||
+          _isExecutingTask) {
+        _cancelDispatchTask();
+      } else {
+        AssistsMessageService.cancelChatTask();
+      }
+
+      setState(() {
+        _isAiResponding = false;
+        _isCheckingExecutableTask = false;
+        _isExecutingTask = false;
+        _isInputAreaVisible = true;
+        _messages.removeWhere(
+          (msg) => msg.isLoading || _isOpenClawWaitingCardMessage(msg),
+        );
+      });
+
+      debugPrint('Task cancelled, all states reset');
+    } catch (e) {
+      debugPrint('onCancelTask error: $e');
+    }
+  }
+
+  @override
+  void _cancelDispatchTask() {
+    final taskId = _currentDispatchTaskId;
+    interruptActiveToolCard();
+    AssistsMessageService.cancelRunningTask();
+    if (taskId != null) {
+      _runtimeCoordinator.unregisterTask(taskId);
+      removeThinkingCard(taskId);
+    }
+    clearAgentStreamSessionState();
+    resetDispatchState();
+  }
+
+  @override
+  void _onCancelTaskFromCard(String taskId) {
+    try {
+      interruptActiveToolCard();
+      if (_isDeepThinking) {
+        AssistsMessageService.cancelRunningTask();
+      }
+      AssistsMessageService.cancelRunningTask();
+      _runtimeCoordinator.unregisterTask(taskId);
+      _updateThinkingCardToCancelled(taskId);
+      clearAgentStreamSessionState();
+      resetDispatchState();
+      setState(() {
+        _isAiResponding = false;
+        _isExecutingTask = false;
+        _isInputAreaVisible = true;
+        _messages.removeWhere(
+          (msg) => msg.isLoading || _isOpenClawWaitingCardMessage(msg),
+        );
+      });
+    } catch (e) {
+      debugPrint('onCancelTaskFromCard error: $e');
+    }
+  }
+
+  @override
+  void _updateThinkingCardToCancelled(String taskId) {
+    final thinkingCards = _messages
+        .where(
+          (msg) =>
+              msg.id == '$taskId-thinking' ||
+              msg.id.startsWith('$taskId-thinking-'),
+        )
+        .toList();
+    if (thinkingCards.isEmpty) return;
+
+    final thinkingCard = thinkingCards.first;
+    final thinkingCardId = thinkingCard.id;
+    final index = _messages.indexWhere((msg) => msg.id == thinkingCardId);
+    if (index == -1) return;
+
+    final cardData = Map<String, dynamic>.from(thinkingCard.cardData ?? {});
+    cardData['stage'] = 5;
+    cardData['isLoading'] = false;
+    cardData['endTime'] = DateTime.now().millisecondsSinceEpoch;
+
+    setState(() {
+      _messages[index] = ChatMessageModel(
+        id: thinkingCardId,
+        type: 2,
+        user: 3,
+        content: {'cardData': cardData, 'id': thinkingCardId},
+      );
+    });
+  }
+
+  @override
+  void _onPopupVisibilityChanged(bool visible) {
+    setState(() {
+      _isPopupVisible = visible;
+    });
+  }
+
+  @override
+  Future<void> _requestAuthorizeForExecution(
+    List<String> requiredPermissionIds,
+  ) async {
+    if (_isAwaitingAuthorizeResult) return;
+    if (latestUserUtterance().trim().isEmpty) return;
+
+    _isAwaitingAuthorizeResult = true;
+    try {
+      final result = await GoRouterManager.pushForResult<bool>(
+        '/home/authorize',
+        extra: AuthorizePageArgs(
+          requiredPermissionIds: requiredPermissionIds.isEmpty
+              ? kTaskExecutionRequiredPermissionIds
+              : requiredPermissionIds,
+        ),
+      );
+      if (result == true && mounted) {
+        await _retryLatestInstructionAfterAuth();
+      }
+    } finally {
+      _isAwaitingAuthorizeResult = false;
+    }
+  }
+
+  @override
+  Future<void> _retryLatestInstructionAfterAuth() async {
+    if (_isRetryingLatestInstructionAfterAuth ||
+        _activeConversationMode == ChatPageMode.openclaw) {
+      return;
+    }
+    if (latestUserUtterance().trim().isEmpty) return;
+
+    _isRetryingLatestInstructionAfterAuth = true;
+    final timestamp = DateTime.now().millisecondsSinceEpoch.toString();
+    final aiMessageId = '$timestamp-ai';
+    final userMessageId = '$timestamp-user';
+
+    try {
+      if (mounted) {
+        setState(() {
+          _removeLatestSystemReplyBeforeAuthRetry();
+          _isAiResponding = true;
+        });
+      }
+
+      final handled = await _handleExecutableTaskFlow(
+        aiMessageId,
+        userMessageId,
+      );
+      if (!handled && mounted && _currentDispatchTaskId == aiMessageId) {
+        handleAgentError('统一 Agent 启动失败，请检查模型提供商与场景模型配置。');
+      }
+    } finally {
+      _isRetryingLatestInstructionAfterAuth = false;
+    }
+  }
+
+  @override
+  void _removeLatestSystemReplyBeforeAuthRetry() {
+    var removeCount = 0;
+    for (final message in _messages) {
+      if (message.user == 1) break;
+      removeCount += 1;
+    }
+    if (removeCount <= 0) return;
+    _messages.removeRange(0, removeCount);
+  }
+}
